sqltypes: move datetime handling to Value method
